+++ GRDB/Statement.swift
@@ -40,10 +40,10 @@ public class Statement {
     public var sql: String
     
     /// The query arguments
-    public var arguments: QueryArguments? {
+    public var arguments: StatementArguments? {
         didSet {
             reset() // necessary before applying new arguments
-            clearQueryArguments()
+            clearArguments()
             if let arguments = arguments {
                 arguments.bindInStatement(self)
             }
@@ -96,7 +96,7 @@ public class Statement {
         }
     }
     
-    // Exposed for QueryArguments. Don't make this one public unless we keep the arguments property in sync.
+    // Exposed for StatementArguments. Don't make this one public unless we keep the arguments property in sync.
     final func bind(value: DatabaseValueConvertible?, atIndex index: Int) {
         let databaseValue = value?.databaseValue ?? .Null
         let code: Int32
@@ -120,7 +120,7 @@ public class Statement {
         }
     }
     
-    // Exposed for QueryArguments. Don't make this one public unless we keep the arguments property in sync.
+    // Exposed for StatementArguments. Don't make this one public unless we keep the arguments property in sync.
     final func bind(value: DatabaseValueConvertible?, forKey key: String) {
         let index = Int(sqlite3_bind_parameter_index(sqliteStatement, ":\(key)"))
         guard index > 0 else {
@@ -138,7 +138,7 @@ public class Statement {
     }
     
     // Don't make this one public or internal unless we keep the arguments property in sync.
-    private func clearQueryArguments() {
+    private func clearArguments() {
         let code = sqlite3_clear_bindings(sqliteStatement)
         if code != SQLITE_OK {
             fatalDatabaseError(DatabaseError(code: code, message: database.lastErrorMessage, sql: sql))